Changes instruction format.
From a solid 4-byte format, we move to a variable width format. This
will help not only in simplifying the instruction set architecture by
way of less encoding issues to deal with, less abuse of registers and
the stack, but also it will help us make our code more compact.

In the common case, instructions will still occupy about 4 bytes given
their operands, but now we treat an instruction as 8-bits. It is then
the responsibility of the instruction to decode its operands. This
offloads responsibility of instruction decoding to the instructions
themselves. I'm not convinced this is a good idea, but we'll go with
it for the time being.

+++ vm/input_reader.cpp
@@ -44,7 +44,7 @@ namespace Caribou
 		file.seekg(0, std::ios::beg);
 
 		machine.allocate_instruction_memory(instruction_size);
-		uint32_t*& instructions = machine.get_instructions();
+		uint8_t*& instructions = machine.get_instructions();
 
 		file.read((char*)&header, sizeof(BytecodeHeader));
 		strncpy(magic, header.name, 4);
